requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
python-docx>=0.8.11
spacy>=3.7.0
scispacy>=0.5.3
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
biopython>=1.81
beautifulsoup4>=4.12.0
nltk>=3.8.1
transformers>=4.30.0
torch>=2.0.0
plotly>=5.15.0
google-generativeai>=0.3.0 